Move download-related code to a separate package
